fix(general ledger): link to journal rather than account stmt
This commit fixes a long broken link in the General Ledger that links to
the account statement report. This report was removed some time ago.  It
replaces this link with the filtered posting journal, which emulates the
same behavior.  It allows users to have a dynamic look at the
transactions affecting the account in question.

Closes #7787.