feat(employees): add a locked search filter
This commit adds a locked filter to the employee search modal.  The
locked filter allows a user to select only employees that have been
locked or unlocked.

I used the bhYesNoRadios to do this, but it required an update to the
parsing logic to allow for undefined values.  I also added the ability
to remove the "required" attribute from the radio buttons so that
  validation wouldn't trigger every submit.

Closes #7798.