chore: clean up bash scripts
This commit removes unused or deprecated scripts in the BHIMA repository
and also improves some of the other bash scripts after passing through
shellcheck linting.  This should relieve some maintenance burden on the
developers.

Maintenance scripts (such as the install script, etc) should be in a
separate repository third-culture-software/bhima-scripts.