chore(rubrics): unify rubric_config server controllers
Migrates the rubric configuration server controllers to use the
`payroll/` prefix in routes.  It also removes the `/setting` suffix from
the routes, opting instead to have an API similar to the cash payments
or invoices API where the items are provided as an array in the main
payment configuration object.

The next step will be to unify the client UI to allow simultaneous
configuration for the payroll rubrics.