chore(payroll): address cosmetic requests
Addresses a number of cosmetic requests discussed in #6097.  Notably,
the English syntax for "nontaxable" and "payslip" have been addressed.
Additionally, the currency exchange rate is not rendered on the payslip
if the currency is the enterprise currency.  However, there is a bug with
the currency rendering on the payslip generator that does not allow the
currency to be properly rendered.  This will be addressed in a future
PR.