refactor(payroll): improve config modal
Complete rewrite of the multiple_payroll payment configuration modal to
be less brittle and provide a better user experience.

I've completely removed the option for the user to set the currency
since no production sites were using it in practice, and it had some
serious bugs.  This reduces the number of imports into the modal and
removes the need to do any currency conversion.

Additionally, I've provided additionally context in the modal about the
employee, their grade, function, number of dependents, etc., to give the
user all the context needed for configuration of the employee.

Closes #7966.