fix(db): change the order of setTimeout()
This is a critical bug.  Because the setTimeout() call returns a promise, the order
of arguments changed.  This means that transactions are never restarted in the database when
a transaction lock occurs.

I'm not sure how to cause a race condition in testing to create transaction locks, otherwise, I would right a unit test.

This is affecting our users in production at Vanga.